# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GYMCO customer API — multi-gym memberships, classes, bookings, check-ins, and fitness tracking.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memberships/sync:
    post:
      operationId: syncMemberMobile
      tags: [memberships]
      summary: Explicitly sync a mobile with YoActiv for the signed-in member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileSyncRequest"
      responses:
        "200":
          description: Mobile saved and server-only sync receipt recorded
          content:
            application/json:
              schema:
                type: object
                required: [synced]
                properties:
                  synced:
                    type: boolean
  /memberships/sync/automatic:
    post:
      operationId: autoSyncMemberMobile
      tags: [memberships]
      summary: Refresh membership using only the signed-in member's server-only sync receipt
      responses:
        "200":
          description: Automatic sync result; confirmation_required means the member must explicitly verify a mobile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomaticMobileSyncResult"
        "503":
          description: The trusted membership could not be refreshed
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
...
components:
  securitySchemes:
    memberBearer:
      type: http
      scheme: bearer
    adminSession:
      type: apiKey
      in: cookie
      name: connect.sid
...
  schemas:
...
    MobileSyncRequest:
      type: object
      required: [mobile]
      properties:
        mobile:
          type: string
    AutomaticMobileSyncResult:
      type: object
      required: [synced, reason]
      properties:
        synced:
          type: boolean
        reason:
          type: string
          enum: [synced, confirmation_required, mobile_conflict]
...